Finally, new!
All new, greatly simplified and more flexible from the GA perspective. ATTiny85 test circuit with a servo on I/O 0 (pin 5) and PIR on I/O 2 (pin 7) was wiggling for the sensor in about 30 minutes. Add an LED on I/O 4 (pin 3, or iLedPin) for minimal diagnostics. Tweak lRootSeed to explore other branches of the gene pool.